SciChart version 2.0.0.486
Revision Message
-------- -------
     483 Merged revision(s) 460-482 from
         branches/SCIGenericType_feasibility_brunch:

     485 Merged revision(s) 483-484 from
         branches/SCIGenericType_feasibility_brunch:

     486 Changed base type in DateValuesFactory to be SCIDataType.DateTime
         instead of Double to fix RealtimeTickingStockChart issue

+++ src/Xamarin.Examples.Demo.iOS/Views/Examples/UsingTooltipModifierTooltipsView.cs
@@ -79,7 +79,7 @@ namespace Xamarin.Examples.Demo.iOS.Views.Examples
             Surface.ChartModifiers = new SCIChartModifierCollection(
                 new SCITooltipModifier
                 {
-                    Style = { HitTestMode = (int)SCIHitTestMode.Interpolate }
+                    Style = { HitTestMode = SCIHitTestMode.Interpolate },
                     //ShowAxisLabels = true,
                     //ShowTooltip = true,
                 }
